Final updates

+++ src/main/java/com/rethink/orders/repositories/OrderRepository.java
@@ -3,6 +3,6 @@ package com.rethink.orders.repositories;
 import com.rethink.orders.models.Order;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Order, Integer> {
+public interface OrderRepository extends CrudRepository<Order, Long> {
 
 }
